templates/python/requirements.txt
# GTM Analysis Module - Python gRPC Template Requirements
# 
# Template placeholders:
# - Replace with your specific version requirements
# - Add any additional dependencies your analyzer needs

# Core gRPC dependencies
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-health-checking>=1.60.0
grpcio-reflection>=1.60.0

# Protocol Buffers
protobuf>=4.25.0

# Data validation and parsing
pydantic>=2.5.0

# Logging and utilities
structlog>=23.2.0

# Testing framework (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-grpc>=0.8.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# TODO: Add your specific analyzer dependencies
# Examples:
# requests>=2.31.0          # For HTTP requests
# beautifulsoup4>=4.12.0    # For HTML parsing
# jsbeautifier>=1.14.0     # For JavaScript formatting
# python-dateutil>=2.8.0   # For date parsing
# regex>=2023.0.0           # For advanced regex patterns